requirements.txt
aiogram~=3.15.0
tortoise-orm~=0.21.7
pydantic-settings~=2.6.0
fluent.runtime~=0.4.0
aiogram_i18n~=1.4
qrcode[pil]~=8.0